m64p core mixed prototypes in header common to 32- and 64-bit drc

+++ mupen64plus-core/src/r4300/x86/gr4300.c
@@ -353,7 +353,7 @@ void gendebug(void)
 }
 #endif
 
-void gencallinterp(unsigned long addr, int jump)
+void gencallinterp(uint32_t addr, int jump)
 {
    free_all_registers();
    simplify_access();
